update: optimized codes, update products
