+++ src/screens/services/products/Products.tsx
@@ -1,6 +1,5 @@
 import { Product } from "./types";
 import { useState } from "react";
-import { useTranslation } from "react-i18next";
 import {
   Carousel,
   CarouselContent,
@@ -14,22 +13,12 @@ import { Label } from "@/components/ui/label";
 import SvgIcons from "@/components/svg-icons/SvgIcons";
 import { cn } from "@/lib/utils";
 import LocalizationKey from "@/i18n/key";
-import { initValue } from "./utils";
-
-import Products12Pro from "@/assets/Products/products/icharge12Pro.png";
-import Products12 from "@/assets/Products/products/icharge12.png";
-import Products8Plus from "@/assets/Products/products/icharge8Plus.png";
-import Products8 from "@/assets/Products/products/icharge8.png";
-import Products6 from "@/assets/Products/products/icharge6.png";
-import Products24Max from "@/assets/Products/products/ichargeMax.png";
-import Products48ProMax from "@/assets/Products/products/iCharge48ProMax.png";
-import ProductsQCRC from "@/assets/Products/products/ichargeQCRC.png";
-import ProductsQCSQ from "@/assets/Products/products/ichargeQCSQ.png";
+import { useServices } from "@/screens/services/utils";
 
 const Features = () => {
-  const { t } = useTranslation();
+  const { productList, t, productInit } = useServices();
   const [open, setOpen] = useState(false);
-  const [content, setContent] = useState<Product>(initValue);
+  const [content, setContent] = useState<Product>(productInit);
   const [currentIndex, setCurrentIndex] = useState(0);
 
   const handleOpen = (e: { stopPropagation: () => void }, item: Product) => {
@@ -45,126 +34,6 @@ const Features = () => {
   const handleCarouselChange = (index: number) => {
     setCurrentIndex(index);
   };
-
-  const products: Product[] = [
-    {
-      productDetails: {
-        name: "iCharge 6",
-        description: t(LocalizationKey.services.icharge6desc),
-        image: Products6,
-        seatCapacity: t(LocalizationKey.services.icharge6desc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "75W", standBy: "3W" },
-        dimensions: { h: "1150mm", l: "260mm", w: "235mm" },
-        content: { code: "QI-6", slots: "6", powerbanks: "4" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge 8",
-        description: t(LocalizationKey.services.icharge8desc),
-        image: Products8,
-        seatCapacity: t(LocalizationKey.services.icharge8desc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "65W", standBy: "3W" },
-        dimensions: { h: "174mm", l: "343mm", w: "188mm" },
-        content: { code: "QI-8", slots: "8", powerbanks: "5" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge 8 Plus",
-        description: t(LocalizationKey.services.icharge8plusdesc),
-        image: Products8Plus,
-        seatCapacity: t(LocalizationKey.services.icharge8plusdesc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "65W", standBy: "3W" },
-        dimensions: { h: "174mm", l: "343mm", w: "188mm" },
-        content: { code: "QI-8Plus", slots: "8", powerbanks: "5" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge 12",
-        description: t(LocalizationKey.services.icharge12desc),
-        image: Products12,
-        seatCapacity: t(LocalizationKey.services.icharge12desc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "75W", standBy: "3W" },
-        dimensions: { h: "260mm", l: "260mm", w: "235mm" },
-        content: { code: "QI-12", slots: "12", powerbanks: "8" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge 12 Pro",
-        description: t(LocalizationKey.services.icharge12prodesc),
-        image: Products12Pro,
-        seatCapacity: t(LocalizationKey.services.icharge12prodesc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "75W", standBy: "3W" },
-        dimensions: { h: "260mm", l: "260mm", w: "235mm" },
-        content: { slots: "12", powerbanks: "8", code: "QI-12Pro" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge 24 Max",
-        description: t(LocalizationKey.services.icharge24maxdesc),
-        image: Products24Max,
-        seatCapacity: t(LocalizationKey.services.icharge24maxdesc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "160W", standBy: "24.2W" },
-        dimensions: { h: "1650mm", l: "540mm", w: "450mm" },
-        content: { code: "QI-Max", slots: "24", powerbanks: "16" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge 48 Pro Max",
-        description: t(LocalizationKey.services.icharge48promaxdesc),
-        image: Products48ProMax,
-        seatCapacity: t(LocalizationKey.services.icharge48promaxdesc2),
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "--W", standBy: "--W" },
-        dimensions: { h: "1650mm", l: "540mm", w: "450mm" },
-        content: { code: "-", slots: "48", powerbanks: "32" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge Standee SR",
-        description: t(LocalizationKey.services.ichargestandeesrdesc),
-        image: ProductsQCRC,
-        seatCapacity: "",
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "-", standBy: "-" },
-        dimensions: { h: "890mm", l: "450mm", w: "288mm" },
-        content: { code: "-", slots: "-", powerbanks: "-" },
-      },
-    },
-    {
-      productDetails: {
-        name: "iCharge Standee SS",
-        description: t(LocalizationKey.services.ichargestandeessdesc),
-        image: ProductsQCSQ,
-        seatCapacity: "",
-      },
-      productDescriptions: {
-        wattage: { fullLoad: "-", standBy: "-" },
-        dimensions: { h: "956mm", l: "360mm", w: "340mm" },
-        content: { code: "-", slots: "-", powerbanks: "-" },
-      },
-    },
-  ];
 
   return (
     <section
@@ -190,7 +59,7 @@ const Features = () => {
           }}
         >
           <CarouselContent>
-            {products.map((data, index) => (
+            {productList.map((data, index) => (
               <CarouselItem
                 key={index}
                 className="custom-lg:basisOneThird flex h-[32rem] w-[95%] justify-center"
@@ -228,7 +97,7 @@ const Features = () => {
                   <div className="flex flex-col items-center px-5">
                     <Label
                       variant="title"
-                      className="flex items-center h-[116px] bg-gradient-to-b from-green-900 to-green-600 bg-clip-text text-center text-transparent"
+                      className="flex h-[116px] items-center bg-gradient-to-b from-green-900 to-green-600 bg-clip-text text-center text-transparent"
                     >
                       {data.productDetails.name.toLocaleUpperCase()}
                     </Label>
@@ -258,7 +127,7 @@ const Features = () => {
         </Carousel>
 
         <div className="custom-lg:hidden mt-5 flex justify-center">
-          {products.map((_, index) => (
+          {productList.map((_, index) => (
             <div
               key={index}
               className={`mx-2 h-3 w-3 cursor-pointer rounded-full ${
